Move offer management table to separate file
