+++ public/locales/en/translation.json
@@ -139,6 +139,9 @@
   "animalType": "Animal type",
   "manage": "Manage",
   "availability": "Availabilty",
+  "editDescription": "Edit description",
+  "editAmenities": "Edit amenities",
+  "editAvailability": "Edit availability",
   "edit": "Edit",
   "delete": "Delete",
   "yes": "Yes",
