+++ public/locales/pl/translation.json
@@ -139,6 +139,9 @@
   "animalType": "Rodzaj zwierzęcia",
   "manage": "Zarządzaj",
   "availability": "Dostępność",
+  "editDescription": "Edytuj opis",
+  "editAmenities": "Edytuj udogodnienia",
+  "editAvailability": "Edytuj dostępność",
   "edit": "Edytuj",
   "delete": "Usuń",
   "yes": "Tak",
